+++ lib/cloudsight.rb
@@ -28,105 +28,9 @@ module Cloudsight
     end
   end
 
-  class Util
-    def self.post(url, params, headers = {})
-      headers['Authorization'] = authorization_header(:post, url, params)
-      RestClient.post(url, params, headers)
-    rescue RestClient::Exception => e
-      e.response
-    end
-
-    def self.get(url, headers = {})
-      headers['Authorization'] = authorization_header(:get, url)
-      RestClient.get(url, headers)
-    rescue RestClient::Exception => e
-      e.response
-    end
-
-    def self.authorization_header(http_method, url, params = {})
-      if Cloudsight.api_key
-        "CloudSight #{Cloudsight.api_key}"
-      else
-        # Exclude image file when generating OAuth header
-        filtered_payload = params.dup
-        filtered_payload.delete('image_request[image]')
-
-        oauth = SimpleOAuth::Header.new(http_method, url, filtered_payload, Cloudsight.oauth_options || {})
-        oauth.to_s
-      end
-    end
-  end
-
-  class Request
-    def self.send(options = {})
-      raise RuntimeError.new("Need to define either oauth_options or api_key") unless Cloudsight.api_key || Cloudsight.oauth_options
-      url = "#{Cloudsight::base_url}/image_requests"
-
-      params = construct_params(options)
-      response = Util.post(url, params)
-      data = JSON.parse(response.body)
-      raise ResponseException.new(data['error']) if data['error']
-      raise UnexpectedResponseException.new(response.body) unless data['token']
-
-      data
-    end
-
-    def self.construct_params(options)
-      params = {}
-      [:locale, :language, :latitude, :longitude, :altitude, :device_id, :ttl].each do |attr|
-        params["image_request[#{attr}]"] = options[attr] if options.has_key?(attr)
-      end
-
-      if options[:focus]
-        params['focus[x]'] = options[:focus][:x]
-        params['focus[y]'] = options[:focus][:y]
-      end
-
-      params['image_request[remote_image_url]'] = options[:url] if options.has_key?(:url)
-      params['image_request[image]'] = options[:file] if options.has_key?(:file)
-      params
-    end
-
-    def self.repost(token, options = {})
-      url = "#{Cloudsight::base_url}/image_requests/#{token}/repost"
-
-      response = Util.post(url, options)
-      return true if response.code == 200 and response.body.to_s.strip.empty?
-
-      data = JSON.parse(response.body)
-      raise ResponseException.new(data['error']) if data['error']
-      raise UnexpectedResponseException.new(response.body) unless data['token']
-
-      data
-    end
-  end
-
-  class Response
-    def self.get(token, options = {})
-      url = "#{Cloudsight::base_url}/image_responses/#{token}"
-
-      response = Util.get(url)
-      data = JSON.parse(response.body)
-      raise ResponseException.new(data['error']) if data['error']
-      raise UnexpectedResponseException.new(response.body) unless data['status']
-
-      data
-    end
-
-    def self.retrieve(token, options = {})
-      options = { poll_wait: 1 }.merge(options)
-
-      data = nil
-      loop do
-        sleep options[:poll_wait]
-        data = Cloudsight::Response.get(token, options)
-        yield data if block_given?
-        break if data['status'] != 'not completed' and data['status'] != 'in progress'
-      end
-
-      data
-    end
-  end
+  require 'cloudsight/api'
+  require 'cloudsight/request'
+  require 'cloudsight/response'
 
   class ResponseException < Exception; end
   class UnexpectedResponseException < Exception; end
